refactor hooks into their own files from Room
